Удалил лишние файлы, разрезал core.php на части, написал тест, пофиксил сразу пару багов.
